# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SISMAP API",
    "description": "API do Sistema de Mapeamento",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/api/room/": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "summary": "Criar uma nova sala",
        "description": "Endpoint para criar uma nova sala no sistema",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Sala criada com sucesso"
          },
          "400": {
            "description": "Erro ao criar sala"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Sala de Reuniões A"
                  },
                  "x": {
                    "type": "integer",
                    "example": 10
                  },
                  "y": {
                    "type": "integer",
                    "example": 20
                  },
                  "description": {
                    "type": "string",
                    "example": "Sala para reuniões pequenas"
                  },
                  "capacity": {
                    "type": "integer",
                    "example": 10
                  },
                  "type": {
                    "type": "string",
                    "example": "meeting"
                  },
                  "floor": {
                    "type": "integer",
                    "example": 1
                  },
                  "building": {
                    "type": "string",
                    "example": "Prédio Principal"
                  },
                  "amenities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "projetor",
                      "quadro branco"
                    ]
                  },
                  "path": {
                    "type": "array",
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    },
                    "example": [
                      [
                        10,
                        10
                      ],
                      [
                        20,
                        15
                      ]
                    ]
                  }
                },
                "required": [
                  "name",
                  "x",
                  "y",
                  "description",
                  "capacity",
                  "type",
                  "floor",
                  "building",
                  "amenities"
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Listar todas as salas",
        "description": "Endpoint para listar todas as salas do sistema",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Salas listadas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "Sala de Reuniões A"
                      },
                      "x": {
                        "type": "integer",
                        "example": 10
                      },
                      "y": {
                        "type": "integer",
                        "example": 20
                      },
                      "description": {
                        "type": "string",
                        "example": "Sala para reuniões pequenas"
                      },
                      "capacity": {
                        "type": "integer",
                        "example": 10
                      },
                      "type": {
                        "type": "string",
                        "example": "meeting"
                      },
                      "floor": {
                        "type": "integer",
                        "example": 1
                      },
                      "building": {
                        "type": "string",
                        "example": "Prédio Principal"
                      },
                      "amenities": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "example": [
                          "projetor",
                          "quadro branco"
                        ]
                      },
                      "path": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "integer"
                          }
                        },
                        "example": [
                          [
                            10,
                            10
                          ],
                          [
                            20,
                            15
                          ]
                        ]
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao buscar salas"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/project/": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Criar um novo projeto",
        "description": "Endpoint para criar um projeto no sistema",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "Novo Projeto"
                    },
                    "type": {
                      "type": "string",
                      "example": "Banner"
                    },
                    "startAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-09-25T10:00:00.000Z"
                    },
                    "endAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-09-25T12:00:00.000Z"
                    },
                    "roomId": {
                      "type": "integer",
                      "example": 1
                    },
                    "room": {
                      "$ref": "#/components/schemas/Room"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao criar projeto"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Novo Projeto"
                  },
                  "type": {
                    "type": "string",
                    "example": "Banner"
                  },
                  "startAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-09-25T10:00:00.000Z"
                  },
                  "endAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-09-25T12:00:00.000Z"
                  },
                  "roomId": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "title",
                  "type",
                  "startAt",
                  "endAt",
                  "roomId"
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Listagem de todos os projetos",
        "description": "Endpoint para listar todos os projetos do sistema",
        "responses": {
          "200": {
            "description": "Projetos listados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "example": "Apresentação de Projeto"
                      },
                      "type": {
                        "type": "string",
                        "example": "palestra"
                      },
                      "startAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-25T10:00:00.000Z"
                      },
                      "endAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-25T12:00:00.000Z"
                      },
                      "roomId": {
                        "type": "integer",
                        "example": 1
                      },
                      "room": {
                        "$ref": "#/components/schemas/Room"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao listar projetos"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para autenticação. Formato: Bearer {token}"
      }
    },
    "schemas": {
      "Room": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Sala de Reuniões A"
                  }
                }
              },
              "x": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "y": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 20
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Sala para reuniões pequenas"
                  }
                }
              },
              "capacity": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "meeting"
                  }
                }
              },
              "floor": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "building": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Prédio Principal"
                  }
                }
              },
              "amenities": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "example": {
                    "type": "array",
                    "example": [
                      "projetor",
                      "quadro branco"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "path": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "array"
                      },
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "integer"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "type": "array",
                    "items": {
                      "type": "array",
                      "example": [
                        20,
                        15
                      ],
                      "items": {
                        "type": "number"
                      }
                    }
                  }
                }
              },
              "createdAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updatedAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  },
                  "nullable": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}